Added Case instruction and made branch no longer use >>=
